+++ changelog.md
@@ -33,4 +33,5 @@
 
 - UX update  
 - background image support  
+- halign, valign, hlock, vlock, backcolor for each oval. Find use example in `contents.xml`